[Hotfix] bbox 실수형으로 변경 및 req dto 누락값 추가

See merge request s14-fintech-finance-sub1/S14P21A406!97

+++ backend/src/main/java/com/zud/backend/common/dto/common/DataField.java
@@ -7,6 +7,6 @@ import lombok.Builder;
 @Builder
 public record DataField<T>(T value, Double confidence, Evidence evidence) {
 	@Builder
-	public record Evidence(Integer pageNum, List<Integer> bbox, String rawText, Double confidence) {
+	public record Evidence(Integer pageNum, List<Double> bbox, String rawText, Double confidence) {
 	}
 }
